refactor(auth): update import paths for consistency and clarity
The import paths for OAuthConfigModule, auth functions, and interfaces are updated to ensure consistency in naming conventions and improve clarity. The changes also align the import paths with the project's directory structure, making it easier to navigate and maintain the codebase.
